Modify mentor_card.dart Finally, make the MentorCard navigate to the detail screen when tapped.
Modify mentor_list_screen.dart Update the MentorListScreen to accept the list of mentors as a parameter instead of defining it internally.

Modify app_routes.dart Open lib/core/app_routes.dart and add a nested route to the /app/mentors path. We'll also need to pass our dummy mentor list to the MentorListScreen to find the correct mentor when a card is tapped.

Signed-off-by: AbdullahRFA <shakibrybmn@gmail.com>