A Skill Barter System:
Instead of posting a "request for help," a user will now post an "exchange offer." This offer will have two sides: what they can give and what they want in return.

A New "Create Exchange" Screen
We'll replace the current CreateRequestScreen with a new form that has two distinct sections:

"I Can Teach...": With fields for offerTitle, offerDescription, and offerTags.

"In Return, I Want to Learn...": With fields for requestTitle, requestDescription, and requestTags.

Signed-off-by: AbdullahRFA <shakibrybmn@gmail.com>